packages/server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alembic==1.16.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Faker==37.1.0
fastapi==0.116.1
fastapi-cli==0.0.7
frozenlist==1.7.0
gotrue==2.12.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
humps==0.2.2
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
packaging==24.2
pillow==11.1.0
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.4
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
sqlmodel==0.0.24
starlette==0.47.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
typer==0.15.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.34.0
uvicorn-worker==0.3.0
watchfiles==1.0.4
websockets==14.0
yarl==1.20.1